backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Data source connectors
boto3==1.34.34
snowflake-connector-python==3.6.0
pandas==2.2.0
numpy==1.26.3
pyarrow==15.0.0

# AI/ML
openai==1.10.0

# Async & Caching
redis==5.0.1
httpx==0.26.0

# Airflow client
apache-airflow-client==2.8.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
